### 2017-12-13 #503 @whedon
[PRE REVIEW]: ivporbit:An R package to estimate the instrumental variables probit model
**Submitting author:** @zedtaha (<a href="http://orcid.org/0000-0002-4488-5774">zaghdoudi Taha</a>)
**Repository:** <a href="https://github.com/zedtaha/ivprobit-1.0" target ="_blank">https://github.com/zedtaha/ivprobit-1.0</a>
**Version:** v1.0
