### 2017-12-13 #503 @arfon
👋 @zedtaha, thanks for your submission to JOSS. Please take a look at the author guidelines and make sure your paper is structured properly please http://joss.theoj.org/about#paper_structure
