### 2017-12-13 949ca50
Update paper.md
