### 2017-12-13 b723049
Update paper.md
