### 2017-12-13 9882d97
Update paper.md
