### 2017-12-13 aba7918
Update paper.md
