### 2017-12-13 897d85a
Update paper.md
