### 2017-12-13 #503 @arfon
Also, @zedtaha, please make sure you address these points in your paper:

- A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience
- A clear statement of need that illustrates the purpose of the software
- A list of key references including a link to the software archive
- Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it

Finally, it looks like your project is not open source. Please make sure you add an open source license to your project repository.
