### 2017-12-15 60f6666
Update paper.md
