### 2017-12-15 2fbb89e
Update paper.bib
