+++ paper/paper.bib
@@ -1,6 +1,11 @@
-% This file was created with JabRef 2.9.2.
-% Encoding: UTF8
 
+@online{zaghdoudi,
+	author = {Haytham M. Fayek},
+	title = {ivporbit:An R package to estimate the instrumental variables probit model},
+	year = 2017,
+	url = {https://github.com/zedtaha/ivprobit-1.0/},
+	urldate = {2017-12-15}
+}
 @ARTICLE{Amemiya1978,
   author = {Amemiya, Takeshi},
   title = {The estimation of a simultaneous equation generalized probit model},
