+++ paper/paper.bib
@@ -1,6 +1,6 @@
 
 @online{zaghdoudi,
-	author = {Haytham M. Fayek},
+	author = {Taha Zaghdoudi},
 	title = {ivporbit:An R package to estimate the instrumental variables probit model},
 	year = 2017,
 	url = {https://github.com/zedtaha/ivprobit-1.0/},
