### 2017-12-15 121e5fb
Update paper.md
