### 2018-02-08 #523 @justinesarey
Hi Thomas,

Got it; I should be able to respond soon.

-JE

--

Dr. Justin Esarey
Associate Professor of Political Science
Rice University
Voice: (678) 383-9629
Fax: (713) 348-5273
E-mail:  <justin@justinesarey.com>justin@justinesarey.com
Web: www.justinesarey.com

On Sun, Feb 4, 2018 at 10:37 AM, Thomas J. Leeper <notifications@github.com>
wrote:

> Hi @justinesarey <https://github.com/justinesarey> just a quick nudge on
> this review!
>
