### 2018-02-20 #523 @justinesarey
I have completed this review. The software appears to work as advertised (at least, it matches results on identical data in Stata 15.1 -- an encouraging sign!). However, the documentation provided in R (when installed using install.packages("ivprobit", dep=T)) makes the software challenging to figure out. The basic issue is that the important information in the help files don't correspond to CRAN practices. Among other issues, (a) the example model should be included in the ivprobit help file and not the overall package help file; (b) the ivprobit and ivprobit.formula sub-methods should be subsumed into a single ivprobit help function (c) there should be actual details of, e.g., what arguments get passed to ivprobit in the ivprobit help file (d) the ivprob function should be hidden or should redirect attention to the main ivprobit function (d) the example data set shouldn't be duplicated three times. I strongly suggest the authors take a look at existing R packages and closely duplicate the structure of their help files.
