### 2018-02-20 #523 @leeper
Thanks, @justinesarey! Very helpful.

@zedtaha, can you address Justin's points here and the incomplete items from the review checklist above? Please ping me here when you've had a chance to respond to all of these points.
