+++ DESCRIPTION
@@ -2,13 +2,12 @@ Package: ivprobit
 Type: Package
 Title: Instrumental variables probit model
 Version: 1.0
-Date: 2017-12-14
-Author: Zaghdoudi Taha
-Maintainer: Zaghdoudi Taha <zedtaha@gmail.com>
-Description: ivprobit fit an Instrumental variables probit model using the
-    generalized least squares estimator
+Date: 2018-02-20
+Author: Taha Zaghdoudi
+Maintainer: Taha Zaghdoudi <zedtaha@gmail.com>
+Description: ivprobit fit an Instrumental variables probit model using the generalized least squares estimator
 License: MIT
-Packaged: 2014-09-25 15:07:02 UTC; Taha
-NeedsCompilation: no
-Repository: CRAN
-Date/Publication: 2014-09-25 22:37:05
+Encoding: UTF-8
+LazyData: true
+RoxygenNote: 6.0.1
+Imports: stats, methods, Formula
