+++ man/eco.Rd
@@ -1,9 +1,0 @@
-\docType{data}
-\name{eco}
-\alias{eco}
-\title{Foreign-Exchange Derivatives Use By Large U.S. Bank Holding Com-panies (1996-2000).}
-\description{
-Foreign-Exchange Derivatives Use By Large U.S. Bank Holding
-Com-panies (1996-2000).
-}
-
