### 2018-02-21 f5ed84e
Delete ivprobit.formula.Rd
