+++ man/ivprobit.formula.Rd
@@ -1,19 +1,0 @@
-\name{ivprobit.formula}
-\alias{ivprobit.formula}
-\title{formula}
-\usage{
-\method{ivprobit}{formula}(formula1, formula2, data = list(), ...)
-}
-\arguments{
-  \item{formula1}{first formula}
-
-  \item{formula2}{second formula}
-
-  \item{data}{the dataset}
-
-  \item{...}{not used}
-}
-\description{
-formula
-}
-
