### 2018-02-21 b07eb3f
Delete mydata.Rd
