### 2018-02-21 06fe931
Update ivprobit-package.Rd
