### 2018-02-21 4a1b6cb
Update ivprobit.Rd
