### 2018-02-21 5f15876
Update summary.ivprobit.Rd
