+++ man/summary.ivprobit.Rd
@@ -1,3 +1,5 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/ivprobit.R
 \name{summary.ivprobit}
 \alias{summary.ivprobit}
 \title{Summary}
@@ -5,9 +7,9 @@
 \method{summary}{ivprobit}(object, ...)
 }
 \arguments{
-  \item{object}{is the object of the function}
+\item{object}{is the object of the function}
 
-  \item{...}{not used}
+\item{...}{not used}
 }
 \description{
 Summary
