### 2018-02-21 25d85bb
Delete ivprobit.R
