+++ NAMESPACE
@@ -1,4 +1,19 @@
-S3method(ivprobit,formula)
+# Generated by roxygen2: do not edit by hand
+
 S3method(summary,ivprobit)
-export(ivprob)
 export(ivprobit)
+import(Formula)
+importFrom(stats,as.formula)
+importFrom(stats,binomial)
+importFrom(stats,coef)
+importFrom(stats,glm)
+importFrom(stats,lm)
+importFrom(stats,model.frame)
+importFrom(stats,model.matrix)
+importFrom(stats,model.response)
+importFrom(stats,na.omit)
+importFrom(stats,pchisq)
+importFrom(stats,printCoefmat)
+importFrom(stats,pt)
+importFrom(stats,update)
+importFrom(stats,vcov)
