### 2018-02-21 #523 @zedtaha
Hi @justinesarey i updated the ivprobit R package close to your comment points. Hope its OK ;)
