### 2018-02-21 #523 @zedtaha
Hi @leeper  i updated my code close to justin points and hope its ok ;) best
