### 2018-02-21 #523 @justinesarey
Is this updated on CRAN as well? I'd like to run
update.packages("ivprobit") and check out the new help files.

-JE

--

Dr. Justin Esarey
Associate Professor of Political Science
Rice University
Voice: (678) 383-9629
Fax: (713) 348-5273
E-mail:  <justin@justinesarey.com>justin@justinesarey.com
Web: www.justinesarey.com

On Wed, Feb 21, 2018 at 8:52 AM, Taha Zaghdoudi <notifications@github.com>
wrote:

> Hi @leeper <https://github.com/leeper> i updated my code close to justin
> points and hope its ok ;) best
>
