### 2018-02-21 #523 @justinesarey
Thanks! I'll take a look once the new package propagates to the CRAN mirror.

-JE

--

Dr. Justin Esarey
Associate Professor of Political Science
Rice University
Voice: (678) 383-9629
Fax: (713) 348-5273
E-mail:  <justin@justinesarey.com>justin@justinesarey.com
Web: www.justinesarey.com

On Wed, Feb 21, 2018 at 9:22 AM, Taha Zaghdoudi <notifications@github.com>
wrote:

> @justinesarey <https://github.com/justinesarey> no just on the git but I
> ll do it now
>
