### 2018-02-21 08cd7fe
Update paper.md
