+++ README.md
@@ -34,4 +34,4 @@ The function is :
  summary(pro)
  ```
 # License
-GPL-3
+All code is licensed [GPL-3](https://www.gnu.org/licenses/gpl-3.0.en.html)
