+++ DESCRIPTION
@@ -1,12 +1,12 @@
 Package: ivprobit
 Type: Package
-Title: Instrumental variables probit model
-Version: 1.0
-Date: 2018-02-20
-Author: Taha Zaghdoudi
-Maintainer: Taha Zaghdoudi <zedtaha@gmail.com>
-Description: ivprobit fit an Instrumental variables probit model using the generalized least squares estimator
-License: MIT
+Title: Instrumental Variables Probit Model
+Version: 1.1
+Date: 2018-02-21
+Author: Zaghdoudi Taha
+Maintainer: Zaghdoudi Taha <zedtaha@gmail.com>
+Description: Compute the instrumental variables probit model using  the Amemiya's Generalized Least Squares estimators (Amemiya, Takeshi, (1978) <doi: 10.2307/1911443>).
+License: GPL-3
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 6.0.1
