### 2018-02-21 23ce7cb
Update ivprobit-package.Rd
