### 2018-02-21 0d7d32c
Update ivprobit-package.Rd
