+++ man/ivprobit-package.Rd
@@ -12,7 +12,7 @@ ivprobit fit an Instrumental variables probit model using the generalized least 
 \tabular{ll}{
 Package: \tab ivprobit\cr
 Type: \tab Package\cr
-Version: \tab 1.0\cr
+Version: \tab 1.1\cr
 Date: \tab 2018-02-20\cr
 License: \tab GPL-3\cr
 }
