### 2018-02-21 39cc4e2
Update ivprobit-package.Rd
