### 2018-02-22 #523 @leeper
@zedtaha Great. Can you please create a DOI for your repository by archiving through [zenodo](https://zenodo.org/), [figshare](https://figshare.com/), or a similar service - and then put the DOI in your README file and as a comment here?
