### 2018-02-22 13237c7
Update paper.md
