+++ README.md
@@ -1,3 +1,8 @@
+---
+output: github_document
+---
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1109726.svg)](https://doi.org/10.5281/zenodo.1109726)
+
 # ivprobit-1.1
 ivprobit fit an Instrumental variables probit model using the generalized least squares estimator
 # Example
@@ -35,3 +40,5 @@ The function is :
  ```
 # License
 All code is licensed [GPL-3](https://www.gnu.org/licenses/gpl-3.0.en.html)
+
+
