### 2018-02-23 #523 @leeper
@zedtaha Great. Once you accept and merge [this Pull Request](https://github.com/zedtaha/ivprobit-1.0/pull/1), then your paper will be accepted at JOSS!
