### 2018-02-23 #523 @leeper
@arfon This one's over to you!
