### 2018-02-23 #523 @arfon
@zedtaha - your paper is now accepted into JOSS and your DOI is https://doi.org/10.21105/joss.00523 ⚡️🚀💥
