### 2018-02-23 #523 @justinesarey
Just FYI: the title of the package is listed as “ivporbit” but I’m pretty
sure the author meant “ivprobit.” Perhaps that can be changed?

-JE

On Fri, Feb 23, 2018 at 8:44 AM Arfon Smith <notifications@github.com>
wrote:

> @justinesarey <https://github.com/justinesarey> - many thanks for your
> review here and to @leeper <https://github.com/leeper> for editing this
> submission ✨
>
> @zedtaha <https://github.com/zedtaha> - your paper is now accepted into
> JOSS and your DOI is https://doi.org/10.21105/joss.00523 ⚡️🚀💥
>
