### 2018-02-23 #523 @arfon
> Just FYI: the title of the package is listed as “ivporbit” but I’m pretty sure the author meant “ivprobit.” Perhaps that can be changed?

Good catch. I've just updated this.
